refactored common dtos, refactored order service

+++ common-dtos/src/main/kotlin/dto/OrderRequestDto.kt
@@ -4,5 +4,5 @@ data class OrderRequestDto(
     var userId: Int,
     var productId: Int,
     var amount: Int,
-    var orderId: Int
+    var orderId: Long
 )
